+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for lambdasound
 
+## Unreleased
+
+* Add `withSampledSound` and `withSampledSoundPulse`
+
 ## 1.0.0 -- 2023-10-12
 
 * First version. Released on an unsuspecting world.